\name{geneNames.CdfEnvAffy}
%\alias{geneNames.AltAffyBatch}
\alias{geneNames.CdfEnvAffy}
\title{ get the names of the known probe sets }
\description{
  get the names of the probe sets known to the CdfEnv
}
\usage{
%geneNames.AltAffyBatch(object)
geneNames.CdfEnvAffy(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{  \code{CdfEnvAffy-class}} %\code{AltAffyBatch-class} or
}
\value{
  a vector of mode \code{character}
}
\keyword{ manip}
