\name{toHypergraph}
\alias{toHypergraph}

\title{Transform to an hypergraph}
\description{
  Transform to an hypergraph
}
\usage{
toHypergraph(object, ...)
}
\arguments{
  \item{object}{Object derived from class \code{AffyProbesMatch}.}
  \item{...}{Unused.}
}
\value{
  An \code{\link[hypergraph]{Hypergraph-class}} object. 
}
\keyword{manip}
